+++ src/views/Profile.vue
@@ -117,6 +117,13 @@
                     {{ $t('profile.view_custom_feeds') }}
                   </router-link>
                 </li>
+                <li v-if="isCurrentUser()">
+                  <router-link
+                    :to="{ name: 'like-list' }"
+                  >
+                    {{ $t('profile.view_likes') }}
+                  </router-link>
+                </li>
                 <li v-if="canViewSubscriber()">
                   <router-link
                     :to="{ name: 'subscriber', params: { profileId: profile.id } }"